first version of AnimalTree with next minor changes: -package renamed; -redundant string removed;
